+++ Source/Core/Memory/FreeListAllocator.h
@@ -4,13 +4,13 @@
 
 STRUCT(FreeListAllocator_Header)
 {
-	u64 BlockSize;
-	u64 AlignmentPadding;
+	usize BlockSize;
+	usize AlignmentPadding;
 };
 
 STRUCT(FreeListAllocator_Node)
 {
-	u64 BlockSize;
+	usize BlockSize;
 	struct FreeListAllocator_Node* Next;
 };
 
@@ -18,8 +18,8 @@ STRUCT(FreeListAllocator)
 {
 	void* Memory;
 	
-	u64 TotalSize;
-	u64 Allocated;
+	usize TotalSize;
+	usize Allocated;
 	
 	FreeListAllocator_Node* Head;
 
@@ -27,13 +27,13 @@ STRUCT(FreeListAllocator)
     struct FreeListAllocator* Next; // pointer to the next memory block
 };
 
-RIFT_API void FreeListAllocator_Create(FreeListAllocator* OutAllocator, u64 TotalSize, void* Memory);
+RIFT_API void FreeListAllocator_Create(FreeListAllocator* OutAllocator, usize TotalSize, void* Memory);
 RIFT_API void FreeListAllocator_Destroy(FreeListAllocator* Allocator);
 
-RIFT_API void* FreeListAllocator_Allocate(FreeListAllocator* Allocator, u64 Size, u64* OutBytesAllocated);
-RIFT_API void FreeListAllocator_Free(FreeListAllocator* Allocator, void* Memory, u64* OutBytesFreed);
+RIFT_API void* FreeListAllocator_Allocate(FreeListAllocator* Allocator, usize Size, usize* OutBytesAllocated);
+RIFT_API void FreeListAllocator_Free(FreeListAllocator* Allocator, void* Memory, usize* OutBytesFreed);
 
 RIFT_API void FreeListAllocator_FreeAll(FreeListAllocator* Allocator);
 
-RIFT_API u64 FreeListAllocator_Offset(FreeListAllocator* Allocator, void* Memory);
-RIFT_API void* FreeListAllocator_MemoryFromOffset(FreeListAllocator* Allocator, u64 Offset);
+RIFT_API usize FreeListAllocator_Offset(FreeListAllocator* Allocator, void* Memory);
+RIFT_API void* FreeListAllocator_MemoryFromOffset(FreeListAllocator* Allocator, usize Offset);
